==> code/analysis/03_analyzing_sampling_completeness_for_experiment.Rmd <==
---
title: "Analyzing sampling completeness"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goals**

The next big step in this analysis is investigating how *Heliconia* removal affected interaction turnoever, i.e. network dissimilarity calculated with bipartite::betalinkr(). Before moving ahead with this analysis, however, I need to explore the extent to which networks were sampled adequately.

Specifically, if pre-to-post sampling completeness differs in treatment replicates - but not control replicates - this systematic bias in sampling might lead to a spurious treatment effect.

To explore this possibility for each sampling method, here I:

1. Summarize sampling completeness
2. Visualize how sampling completeness varies pre-to-post, control-vs-treatment
3. Statistically test whether sampling completeness varies pre-to-post, control-vs-treatment

I conclude that:

1. Sampling completeness was higher for visitation (camera) networks, which were sampled through an array of trail cameras.
2. For pollen networks, pre-to-post sampling completeness was lower in treatment replicates than in control replicates, introducing systematic sampling bias. **Therefore, for the analysis of interaction turnover, I will only analyze camera networks.**

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(dplyr)
library(tidyr)
library(ggplot2)
library(glmmTMB)
library(emmeans)
```

## Step 1: Get data

```{r step1}
#Data
completeness <- read.csv("../../data/export/for_analysis/Sampling_completeness_for_analysis.csv") %>%
  mutate(year_patch = paste(year, patch, sep = "_")) %>%
  mutate(outliers = "with_outliers")

#Add a dataset without outliers
completeness_with_without_outliers <- completeness %>%
  filter(sampling_method == "visitation") %>%
  filter(year_patch != "2017_130") %>%
  filter(year_patch != "2017_24") %>%
  mutate(outliers = "without_outliers") %>%
  bind_rows(completeness)

#Functions
source("../../code/helper_functions/Calculate_basic_summary_stats.R")
source("../../code/helper_functions/Transform_proportion_for_betareg.R")
```

## Step 2: Summarize sampling completeness
```{r step2, message = FALSE}
completeness_sum <- completeness_with_without_outliers %>%
  group_by(sampling_method, outliers) %>%
  calculate_basic_summary_stats(variable = completeness.chao1) %>% 
  ungroup() %>%
  mutate_if(is.numeric, ~round(., digits = 2))

completeness_sum
```

## Step 3: Visualize data
```{r step3, message = FALSE, fig.width = 7, fig.height = 7}
#Prepare data for plotting, including adding indicator variable for outliers
data_for_plot <- completeness_with_without_outliers %>%
  filter(outliers == "with_outliers") %>%
  mutate(outlier_indicator = ifelse(sampling_method == "visitation" & (year_patch == "2017_130" | year_patch == "2017_24"), "outlier", "not_outlier")) %>%
  mutate(control_treatment = factor(control_treatment, levels = c("control", "treatment"), labels = c("Control", "Treatment"))) %>%
  mutate(exp_phase = factor(exp_phase, levels = c("pre", "post"), labels = c("Pre", "Post"))) %>%
  mutate(sampling_method = factor(sampling_method, levels = c("pollen", "visitation"), labels = c("Pollen", "Camera")))

#Make plot
colors <- c("#0E0D37", "#BA0022") #dark blue, dark red

chao1_plot <- data_for_plot %>%
  ggplot(data = ., aes(x = exp_phase, y = completeness.chao1)) +
    facet_grid(control_treatment ~ sampling_method) +
    geom_boxplot(aes(fill = control_treatment), width = 0.15, position = position_nudge(x = c(-0.15, 0.15)), alpha = 0.5, outlier.shape = NA) +
    geom_point(aes(fill = control_treatment), colour = "black", shape = 21, position = position_dodge(width = 0.45), alpha = 1, size = 2) +
    stat_summary(fun = mean, position = position_nudge(x = c(-0.15, 0.15)), geom = "point", shape = 18, size = 3) +
    geom_line(aes(group = year_patch, linetype = outlier_indicator), alpha = 0.6, linewidth = 0.5) +
    theme_bw(base_size = 20) +
    scale_fill_manual(values = colors) +
    scale_color_manual(values = colors) +
    scale_linetype_discrete(c("solid", "dashed")) +
    theme(legend.position = "none",
          legend.direction = "horizontal",
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          strip.background = element_blank(),
          panel.border = element_rect(colour = "black", fill = NA))+
    labs(x = "", y = "Sampling completeness (Chao 1)", fill = "", colour = "") +
    scale_y_continuous(limits = c(0, 1))
    
chao1_plot
```

## Step 4: Statistically analyze pre-to-post changes in sampling completeness. Are there greater changes in sampling completeness depending on treatment?

Transform response variable so that it falls between 0 and 1, so I can use beta regression
```{r step4a, message = FALSE}
completeness_for_analysis <- completeness_with_without_outliers %>%
  group_by(sampling_method, outliers) %>%
  mutate(chao1_transform = transform_proportion(completeness.chao1)) %>%
  ungroup() %>%
  mutate(exp_phase = factor(exp_phase, levels = c("pre", "post"))) 
```

Create model for examine for each sampling method and check assumptions
```{r step4b, message = FALSE,  fig.show = 'hide', results = 'hide'}
model_pollen <- glmmTMB::glmmTMB(chao1_transform ~ exp_phase * control_treatment + (1|patch/year_patch), data = filter(completeness_for_analysis, sampling_method == "pollen"), family = "beta_family")

model_camera <- glmmTMB::glmmTMB(chao1_transform ~ exp_phase * control_treatment + (1|patch/year_patch), data = filter(completeness_for_analysis, sampling_method == "visitation" & outliers == "with_outliers"), family = "beta_family")

model_camera_without_outliers <- glmmTMB::glmmTMB(chao1_transform ~ exp_phase * control_treatment + (1|patch/year_patch), data = filter(completeness_for_analysis, sampling_method == "visitation" & outliers == "without_outliers"), family = "beta_family")

DHARMa::simulateResiduals(model_pollen) %>% plot()
DHARMa::simulateResiduals(model_camera) %>% plot()
DHARMa::simulateResiduals(model_camera_without_outliers) %>% plot()
```

Calculate relevant contrasts
```{r step4c, message = FALSE}
#Function to calculate beta regression contrasts as probability ratios, as in script 05
calculate_contrasts <- function(model){
  
  emm = emmeans(model, specs = ~exp_phase|control_treatment) 
  emm_log = regrid(emm, transform = "log")
  #result = summary(contrast(emm_log, interaction = "revpairwise"), type = "response") #Without CI, with p-value
  result = confint(contrast(emm_log, interaction = "revpairwise"), type = "response") #With CI, without p-value
  
  return(result)
  
}

completeness_contrasts <- tibble(model_number = c(1, 2, 3),
                                 model_name = c("Pollen", "Camera", "Camera (without outliers)"),
                                 model = list(model_pollen, model_camera, model_camera_without_outliers)) %>%
  mutate(contrasts = purrr::map(model, calculate_contrasts)) %>%
  unnest(contrasts) %>%
  mutate_if(is.numeric, ~round(., digits = 2)) %>%
  select(-model) %>% 
  select(model_number, model_name, exp_phase_revpairwise, control_treatment, ratio, lower.CL, upper.CL, everything())
```

## Step 5: Export results
```{r step5}
path_name <- c("../../results/analyses_sampling_completeness")
write.csv(completeness_sum, file.path(path_name, "tables/Sampling_completeness_summary.csv"), row.names = FALSE)
write.csv(completeness_contrasts, file.path(path_name, "tables/Sampling_completeness_contrasts.csv"), row.names = FALSE)
ggsave(file.path(path_name, "figures/Sampling_completeness_plot.png"), plot = chao1_plot, width = 7, height = 7, units = "in", dpi = 600)
```
